search-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
numpy==1.24.3
faiss-cpu==1.7.4
torch==2.1.0
torchvision==0.16.0
pillow==10.1.0
ftfy==6.1.1
regex==2023.10.3
tqdm==4.66.1
git+https://github.com/openai/CLIP.git
